+++ src/Promitor.Scraper.Host/Models/HealthMonitor.cs
@@ -8,10 +8,13 @@ namespace Promitor.Scraper.Host.Models
     public class HealthMonitor
     {
         public int subscriptionLimitCount { get; set; }
-        public HealthMonitor()
+        private HealthMonitor()
         {
             subscriptionLimitCount = int.MinValue;
         }
+
+        private static Lazy<HealthMonitor> _instance = new Lazy<HealthMonitor>(() => new HealthMonitor());
+        public static HealthMonitor Instance { get; } = _instance.Value;
 
     }
 }
